recon_tool/requirements.txt
dnspython
requests
groq
python-dotenv